Well this is less ugly at least
