+++ include/ccf/service/node_info.h
@@ -2,20 +2,15 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/crypto/key_pair.h"
-#include "ccf/crypto/san.h"
-#include "ccf/crypto/sha256_hash.h"
-#include "ccf/crypto/verifier.h"
-#include "ccf/entity_id.h"
-#include "ccf/kv/map.h"
-#include "ccf/kv/serialisers/blit_serialiser_entity_id.h"
-#include "ccf/kv/serialisers/blit_serialiser_pem.h"
+#include "ccf/crypto/pem.h"
+#include "ccf/ds/json.h"
+#include "ccf/kv/version.h"
 #include "ccf/quote_info.h"
-#include "ccf/service/map.h"
-#include "node_info_network.h"
+#include "ccf/service/node_info_network.h"
 
-#include <string>
-#include <vector>
+#define FMT_HEADER_ONLY
+#include <fmt/format.h>
+#include <optional>
 
 namespace ccf
 {
@@ -34,10 +29,7 @@ namespace ccf
      {NodeStatus::RETIRED, "Retired"},
      {NodeStatus::LEARNER, "Learner"},
      {NodeStatus::RETIRING, "Retiring"}});
-}
 
-namespace ccf
-{
   struct NodeInfo : NodeInfoNetwork
   {
     /// Node enclave quote
@@ -84,35 +76,6 @@ namespace ccf
     code_digest,
     certificate_signing_request,
     public_key);
-
-  using Nodes = ServiceMap<NodeId, NodeInfo>;
-  using NodeEndorsedCertificates =
-    kv::RawCopySerialisedMap<NodeId, crypto::Pem>;
-
-  namespace Tables
-  {
-    static constexpr auto NODES = "public:ccf.gov.nodes.info";
-    static constexpr auto NODE_ENDORSED_CERTIFICATES =
-      "public:ccf.gov.nodes.endorsed_certificates";
-  }
-
-  inline NodeId compute_node_id_from_pubk_der(
-    const std::vector<uint8_t>& node_pubk_der)
-  {
-    return crypto::Sha256Hash(node_pubk_der).hex_str();
-  }
-
-  inline NodeId compute_node_id_from_cert_der(
-    const std::vector<uint8_t>& node_cert_der)
-  {
-    return compute_node_id_from_pubk_der(
-      crypto::public_key_der_from_cert(node_cert_der));
-  }
-
-  inline NodeId compute_node_id_from_kp(const crypto::KeyPairPtr& node_sign_kp)
-  {
-    return compute_node_id_from_pubk_der(node_sign_kp->public_key_der());
-  }
 }
 
 FMT_BEGIN_NAMESPACE
