+++ src/ds/cli_helper.h
@@ -4,7 +4,7 @@
 
 #include "ccf/crypto/san.h"
 #include "ccf/ds/nonstd.h"
-#include "service/tables/node_info_network.h"
+#include "ccf/service/node_info_network.h"
 
 #include <CLI11/CLI11.hpp>
 #include <charconv>
