+++ src/endpoints/authentication/cert_auth.cpp
@@ -5,8 +5,8 @@
 
 #include "ccf/rpc_context.h"
 #include "ccf/service/tables/members.h"
+#include "ccf/service/tables/nodes.h"
 #include "ccf/service/tables/users.h"
-#include "service/tables/nodes.h"
 
 namespace ccf
 {
