+++ src/node/hooks.h
@@ -3,9 +3,9 @@
 #pragma once
 
 #include "ccf/ds/logger.h"
+#include "ccf/service/node_info_network.h"
 #include "enclave/reconfiguration_type.h"
 #include "service/tables/config.h"
-#include "service/tables/node_info_network.h"
 #include "service/tables/signatures.h"
 
 namespace ccf
