+++ src/node/rpc/call_types.h
@@ -4,6 +4,7 @@
 
 #include "ccf/ds/json_schema.h"
 #include "ccf/receipt.h"
+#include "ccf/service/tables/nodes.h"
 #include "ccf/tx_id.h"
 #include "ccf/tx_status.h"
 #include "kv/kv_types.h"
@@ -11,7 +12,6 @@
 #include "node/ledger_secrets.h"
 #include "node_call_types.h"
 #include "service/tables/code_id.h"
-#include "service/tables/nodes.h"
 #include "service/tables/service.h"
 
 #include <nlohmann/json.hpp>
