+++ src/node/rpc/node_call_types.h
@@ -4,6 +4,7 @@
 
 #include "ccf/ds/json_schema.h"
 #include "ccf/node_startup_state.h"
+#include "ccf/service/node_info_network.h"
 #include "ccf/service/tables/members.h"
 #include "common/configuration.h"
 #include "enclave/interface.h"
@@ -11,7 +12,6 @@
 #include "node/ledger_secrets.h"
 #include "service/tables/code_id.h"
 #include "service/tables/config.h"
-#include "service/tables/node_info_network.h"
 #include "service/tables/service.h"
 
 #include <nlohmann/json.hpp>
