+++ src/node/test/snapshot.cpp
@@ -2,10 +2,10 @@
 // Licensed under the Apache 2.0 License.
 
 #include "ccf/crypto/key_pair.h"
+#include "ccf/service/tables/nodes.h"
 #include "kv/test/null_encryptor.h"
 #include "kv/test/stub_consensus.h"
 #include "node/history.h"
-#include "service/tables/nodes.h"
 #include "service/tables/signatures.h"
 
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
