+++ src/service/genesis_gen.h
@@ -4,11 +4,11 @@
 
 #include "ccf/crypto/verifier.h"
 #include "ccf/service/tables/members.h"
+#include "ccf/service/tables/nodes.h"
 #include "ccf/tx.h"
 #include "network_tables.h"
 #include "node/ledger_secrets.h"
 #include "service/tables/code_id.h"
-#include "service/tables/nodes.h"
 
 #include <algorithm>
 #include <fstream>
