+++ src/service/tables/code_id.h
@@ -2,9 +2,8 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "service/tables/code_digest.h"
-#include "ccf/kv/serialisers/blit_serialiser_code_digest.h"
 #include "ccf/service/map.h"
+#include "service/tables/code_digest.h"
 
 namespace ccf
 {
@@ -13,4 +12,24 @@ namespace ccf
   {
     static constexpr auto NODE_CODE_IDS = "public:ccf.gov.nodes.code_ids";
   }
+}
+
+namespace kv::serialisers
+{
+  template <>
+  struct BlitSerialiser<ccf::CodeDigest>
+  {
+    static SerialisedEntry to_serialised(const ccf::CodeDigest& code_digest)
+    {
+      auto hex_str = ds::to_hex(code_digest.data);
+      return SerialisedEntry(hex_str.begin(), hex_str.end());
+    }
+
+    static ccf::CodeDigest from_serialised(const SerialisedEntry& data)
+    {
+      ccf::CodeDigest ret;
+      ds::from_hex(std::string(data.data(), data.end()), ret.data);
+      return ret;
+    }
+  };
 }
